Refactor product DTOs and update controller methods
Updated `ProductsController` to use `CreateProductDto` and added a new `Put` method for product updates. Created `CreateProductDto` and `UpdateProductDto` classes with validation attributes. Modified `ProductDto` to remove certain validations and added a `LastModifiedOn` property. Updated mapping in `ProductMappingProfile` and changed method signatures in `IProductService` and `ProductService` to accommodate the new DTOs.